Remove deprecated MavenRuntime and IMavenLauncherConfiguration API
