+++ org.eclipse.m2e.core/META-INF/MANIFEST.MF
@@ -28,7 +28,7 @@ Export-Package: org.eclipse.m2e.core,
  org.eclipse.m2e.core.internal.index;x-internal:=true,
  org.eclipse.m2e.core.internal.index.filter;x-internal:=true,
  org.eclipse.m2e.core.internal.jobs;x-internal:=true,
- org.eclipse.m2e.core.internal.launch;x-internal:=true,
+ org.eclipse.m2e.core.internal.launch;x-friends:="org.eclipse.m2e.launching",
  org.eclipse.m2e.core.internal.lifecyclemapping;x-internal:=true,
  org.eclipse.m2e.core.internal.lifecyclemapping.discovery;x-internal:=true,
  org.eclipse.m2e.core.internal.lifecyclemapping.model;x-internal:=true,
